# docs/openapi.yaml
openapi: 3.0.3
info:
  title: oServ API (Next.js)
  version: "1.0.0"
  description: |
    Documentação da API do sistema **oServ** após a migração para **Next.js Fullstack**.  
    Todas as rotas de API estão sob o prefixo `/api` e seguem a arquitetura multiempresa descrita nos documentos do projeto.  
    As principais rotas listadas no arquivo `README.md` incluem autenticação, gerenciamento da empresa, perfis, cadastros, ordens de serviço, comentários e notificações【890973672335846†L291-L352】.  
    Esta especificação também considera as entidades e campos sugeridos na proposta de arquitetura SaaS multiempresa【169360353382499†L31-L120】.
servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento
paths:
...
  /api/auth/register-company:
    post:
      summary: Registrar empresa e usuário proprietário
      description: |
        Cria uma nova empresa assinante, cadastra o usuário **OWNER** e inicia a assinatura inicial.  
        O endpoint é listado como rota principal para autenticação【890973672335846†L291-L299】 e recomendado para o fluxo de primeiro registro【169360353382499†L152-L160】.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterCompanyRequest'
      responses:
        "201":
          description: Empresa criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterCompanyResponse'
        "400":
          description: Dados inválidos ou subdomínio indisponível
  /api/auth/login:
    post:
      summary: Autenticar usuário
      description: |
        Efetua o login de um usuário existente. Retorna um token de autenticação (cookie HTTP-only) e redirecionamento para o dashboard apropriado.  
        Rota listada na seção de autenticação【890973672335846†L291-L299】.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "401":
          description: Credenciais inválidas
...
  /api/company:
    get:
      summary: Obter dados da empresa atual
      description: |
        Retorna as informações da empresa autenticada.  
        Rota citada em `README.md`【890973672335846†L302-L308】.
      responses:
        "200":
          description: Dados da empresa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        "401":
          description: Não autenticado
    patch:
      summary: Atualizar dados da empresa
      description: |
        Atualiza os dados da empresa atual.  
        Os campos enviados no corpo substituirão os valores existentes.  
        Rota citada em `README.md`【890973672335846†L302-L308】.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyUpdateRequest'
      responses:
        "200":
          description: Empresa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        "400":
          description: Dados inválidos
        "401":
          description: Não autenticado
...
  /api/customers:
    get:
      summary: Listar clientes
      description: |
        Retorna todos os clientes cadastrados da empresa autenticada.  
        Endpoint listado em `README.md`【890973672335846†L310-L327】.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
          description: Página para paginação (opcional).
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
          description: Quantidade de itens por página (opcional).
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
        "401":
          description: Não autenticado
    post:
      summary: Cadastrar cliente
      description: |
        Cria um novo cliente para a empresa autenticada.  
        Endpoint listado em `README.md`【890973672335846†L310-L327】.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreateRequest'
      responses:
        "201":
          description: Cliente criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        "400":
          description: Dados inválidos
        "401":
          description: Não autenticado
    patch:
      summary: Atualizar cliente
      description: |
        Atualiza os dados de um cliente existente. O corpo deve conter o identificador do cliente e os campos a serem atualizados.  
        Endpoint listado em `README.md`【890973672335846†L310-L327】.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateRequest'
      responses:
        "200":
          description: Cliente atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        "400":
          description: Dados inválidos
        "401":
          description: Não autenticado
    delete:
      summary: Excluir cliente
      description: |
        Remove um cliente pelo identificador informado na query string.  
        Endpoint listado em `README.md`【890973672335846†L310-L327】.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: Identificador do cliente a ser removido.
      responses:
        "204":
          description: Cliente excluído com sucesso
        "401":
          description: Não autenticado
...
components:
  schemas:
...
    PlanId:
      type: string
      description: Plano de assinatura escolhido.
      enum:
        - STARTER
        - PRO
        - ENTERPRISE
    Role:
      type: string
      description: Papel do usuário no sistema【169360353382499†L25-L29】.
      enum:
        - OWNER
        - ADMIN
        - EMPLOYEE
        - CUSTOMER
    Status:
      type: string
      description: Status genérico (ativo/inativo).
      enum:
        - ACTIVE
        - INACTIVE
...
    Company:
      type: object
      description: Entidade empresa assinante【169360353382499†L33-L47】.
      properties:
        id:
          type: string
        name:
          type: string
        tradeName:
          type: string
        document:
          type: string
          description: CNPJ ou CPF.
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          $ref: '#/components/schemas/Status'
        subdomain:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - document
        - subdomain
        - status
    User:
      type: object
      description: Entidade usuário【169360353382499†L49-L64】.
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/Status'
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - name
        - email
        - role
        - status
    Customer:
      type: object
      description: Entidade cliente【169360353382499†L66-L80】.
      properties:
        id:
          type: string
        companyId:
          type: string
        userId:
          type: string
          nullable: true
        name:
          type: string
        document:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - name
...
    RegisterCompanyRequest:
      type: object
      description: Dados necessários para registrar uma nova empresa e seu proprietário.
      properties:
        ownerName:
          type: string
        ownerEmail:
          type: string
          format: email
        ownerPassword:
          type: string
        companyName:
          type: string
        tradeName:
          type: string
        document:
          type: string
        companyEmail:
          type: string
          format: email
        companyPhone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        subdomain:
          type: string
        plan:
          $ref: '#/components/schemas/PlanId'
        rememberMe:
          type: boolean
      required:
        - ownerName
        - ownerEmail
        - ownerPassword
        - companyName
        - document
        - subdomain
        - plan
    RegisterCompanyResponse:
      type: object
      description: Resposta ao registrar empresa.
      properties:
        company:
          $ref: '#/components/schemas/Company'
        owner:
          $ref: '#/components/schemas/User'
        redirectTo:
          type: string
          description: URL de redirecionamento após o cadastro (ex. `/dashboard`).
    LoginRequest:
      type: object
      description: Dados necessários para efetuar login.
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        rememberMe:
          type: boolean
      required:
        - email
        - password
    LoginResponse:
      type: object
      description: Resposta de login.
      properties:
        redirectTo:
          type: string
          description: URL de redirecionamento após login.
    CompanyUpdateRequest:
      type: object
      description: Campos permitidos para atualizar a empresa.
      properties:
        name:
          type: string
        tradeName:
          type: string
        document:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          $ref: '#/components/schemas/Status'
        subdomain:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
...
    CustomerCreateRequest:
      type: object
      description: Dados necessários para criar um cliente.
      properties:
        name:
          type: string
        document:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
      required:
        - name
    CustomerUpdateRequest:
      type: object
      description: Campos permitidos para atualizar um cliente.
      properties:
        id:
          type: string
        name:
          type: string
        document:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
...